Drop the fragment for the fetched schema URL
Looks like tv4 changed, and in versions ^1.1 it started treating the URL-s with
fragments differently, causing the following result:

```
$ echo '[]' | ./scripts/json-validate.js 'https://www.swisshotels.com/ota/schema#/definitions/availableHotel'
VALID.
```

`echo '{}'` would yield the same false-positive. Moreover, one could pass
anything in the `#` fragment of the type ID, and it'd still be reported as
`VALID.`

This change restores the actual validation happening for the `#/definitions/*`
types.